Silence DEPRECATION WARNING due to eager loading
The .includes query method is meant for the eager loading of
associations. As of Rails 4, adding a string condition of the
included association results in a deprecation warning.

Add ActiveRecord::QueryMethods.references to "app/models/post.rb"
by passing ".references(:content)" which is a mandatory for
every post being submitted.