Enable bury configuration for User "vagrant".
As we already have a static, known IP Address
set up for the Vagrant-machine we should ship
a bury configuration file for the default
User vagrant too.

Set "action" on creating files via templates to
"create_if_missing" instead of "create", which
is the default to prevent existing files to be
overridden.

Adapt README.md as the sudo(8) call was dropped
accidentally.